backend/requirements-clean.txt
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
python-dotenv==0.21.0
gunicorn==22.0.0